Add AlterReadwriteSplittingRuleStatementUpdater
